adding code for paralleism of make features
