+++ egs/iam/v1/local/make_features.py
@@ -2,7 +2,6 @@
 
 # Copyright      2017  Chun Chieh Chang
 #                2017  Ashish Arora
-#                2017  Yiwen Shao
 #                2018  Hossein Hadian
 
 """ This script converts images to Kaldi-format feature matrices. The input to
@@ -22,14 +21,8 @@
 import argparse
 import os
 import sys
-import scipy.io as sio
 import numpy as np
-import random
 from scipy import misc
-from scipy.ndimage.interpolation import affine_transform
-import math
-from signal import signal, SIGPIPE, SIG_DFL
-signal(SIGPIPE, SIG_DFL)
 
 parser = argparse.ArgumentParser(description="""Converts images (in 'dir'/images.scp) to features and
                                                 writes them to standard output in text format.""")
@@ -42,10 +35,7 @@ parser.add_argument('--feat-dim', type=int, default=40,
 parser.add_argument('--padding', type=int, default=5,
                     help='Number of white pixels to pad on the left'
                     'and right side of the image.')
-parser.add_argument('--vertical-shift', type=int, default=10,
-                    help='total number of padding pixel per column')
-parser.add_argument('--horizontal-shear', type=int, default=45,
-                    help='maximum horizontal shearing degree')
+
 
 args = parser.parse_args()
 
@@ -66,23 +56,18 @@ def write_kaldi_matrix(file_handle, matrix, key):
             file_handle.write("\n")
     file_handle.write(" ]\n")
 
-
-def get_scaled_image(im):
+def get_scaled_image(im, allowed_lengths = None):
     scale_size = args.feat_dim
-    sx = im.shape[1]  # width
-    sy = im.shape[0]  # height
+    sx = im.shape[1]
+    sy = im.shape[0]
     scale = (1.0 * scale_size) / sy
     nx = int(scale_size)
     ny = int(scale * sx)
     im = misc.imresize(im, (nx, ny))
-    return im
-
-
-def horizontal_pad(im, allowed_lengths = None):
     if allowed_lengths is None:
         left_padding = right_padding = args.padding
     else:  # Find an allowed length for the image
-        imlen = im.shape[1] # width
+        imlen = im.shape[1]
         allowed_len = 0
         for l in allowed_lengths:
             if l > imlen:
@@ -92,9 +77,9 @@ def horizontal_pad(im, allowed_lengths = None):
             #  No allowed length was found for the image (the image is too long)
             return None
         padding = allowed_len - imlen
-        left_padding = int(padding // 2)
+        left_padding = padding // 2
         right_padding = padding - left_padding
-    dim_y = im.shape[0] # height
+    dim_y = im.shape[0]
     im_pad = np.concatenate((255 * np.ones((dim_y, left_padding),
                                            dtype=int), im), axis=1)
     im_pad1 = np.concatenate((im_pad, 255 * np.ones((dim_y, right_padding),
@@ -104,145 +89,10 @@ def horizontal_pad(im, allowed_lengths = None):
 ### main ###
 data_list_path = os.path.join(args.dir, 'images.scp')
 
-def contrast_normalization(im, low_pct, high_pct):
-    element_number = im.size
-    rows = im.shape[0]
-    cols = im.shape[1]
-    im_contrast = np.zeros(shape=im.shape)
-    low_index = int(low_pct * element_number)
-    high_index = int(high_pct * element_number)
-    sorted_im = np.sort(im, axis=None)
-    low_thred = sorted_im[low_index]
-    high_thred = sorted_im[high_index]
-    for i in range(rows):
-        for j in range(cols):
-            if im[i, j] > high_thred:
-                im_contrast[i, j] = 255  # lightest to white
-            elif im[i, j] < low_thred:
-                im_contrast[i, j] = 0  # darkest to black
-            else:
-                # linear normalization
-                im_contrast[i, j] = (im[i, j] - low_thred) * \
-                    255 / (high_thred - low_thred)
-    return im_contrast
-
-
-def geometric_moment(frame, p, q):
-    m = 0
-    for i in range(frame.shape[1]):
-        for j in range(frame.shape[0]):
-            m += (i ** p) * (j ** q) * frame[i][i]
-    return m
-
-
-def central_moment(frame, p, q):
-    u = 0
-    x_bar = geometric_moment(frame, 1, 0) / \
-        geometric_moment(frame, 0, 0)  # m10/m00
-    y_bar = geometric_moment(frame, 0, 1) / \
-        geometric_moment(frame, 0, 0)  # m01/m00
-    for i in range(frame.shape[1]):
-        for j in range(frame.shape[0]):
-            u += ((i - x_bar)**p) * ((j - y_bar)**q) * frame[i][j]
-    return u
-
-
-def height_normalization(frame, w, h):
-    frame_normalized = np.zeros(shape=(h, w))
-    alpha = 4
-    x_bar = geometric_moment(frame, 1, 0) / \
-        geometric_moment(frame, 0, 0)  # m10/m00
-    y_bar = geometric_moment(frame, 0, 1) / \
-        geometric_moment(frame, 0, 0)  # m01/m00
-    sigma_x = (alpha * ((central_moment(frame, 2, 0) /
-                         geometric_moment(frame, 0, 0)) ** .5))  # alpha * sqrt(u20/m00)
-    sigma_y = (alpha * ((central_moment(frame, 0, 2) /
-                         geometric_moment(frame, 0, 0)) ** .5))  # alpha * sqrt(u02/m00)
-    for x in range(w):
-        for y in range(h):
-            i = int((x / w - 0.5) * sigma_x + x_bar)
-            j = int((y / h - 0.5) * sigma_y + y_bar)
-            frame_normalized[x][y] = frame[i][j]
-    return frame_normalized
-
-
-def find_slant_project(im):
-    rows = im.shape[0]
-    cols = im.shape[1]
-    std_max = 0
-    alpha_max = 0
-    col_disp = np.zeros(90, int)
-    proj = np.zeros(shape=(90, cols + 2 * rows), dtype=int)
-    for r in range(rows):
-        for alpha in range(-45, 45, 1):
-            col_disp[alpha] = int(r * math.tan(alpha / 180.0 * math.pi))
-        for c in range(cols):
-            if im[r, c] < 100:
-                for alpha in range(-45, 45, 1):
-                    proj[alpha + 45, c + col_disp[alpha] + rows] += 1
-    for alpha in range(-45, 45, 1):
-        proj_histogram, bin_array = np.histogram(proj[alpha + 45, :], bins=10)
-        proj_std = np.std(proj_histogram)
-        if proj_std > std_max:
-            std_max = proj_std
-            alpha_max = alpha
-    proj_std = np.std(proj, axis=1)
-    return -alpha_max
-
-
-def horizontal_shear(im, degree):
-    rad = degree / 180.0 * math.pi
-    padding_x = int(abs(np.tan(rad)) * im.shape[0])
-    padding_y = im.shape[0]
-    if rad > 0:
-        im_pad = np.concatenate(
-            (255 * np.ones((padding_y, padding_x), dtype=int), im), axis=1)
-    elif rad < 0:
-        im_pad = np.concatenate(
-            (im, 255 * np.ones((padding_y, padding_x), dtype=int)), axis=1)
-    else:
-        im_pad = im
-    shear_matrix = np.array([[1, 0],
-                             [np.tan(rad), 1]])
-    # sheared_im = affine_transform(image, shear_matrix, output_shape=(
-    # im.shape[0], im.shape[1] + abs(int(im.shape[0] * np.tan(shear)))), cval=128.0)
-    sheared_im = affine_transform(im_pad, shear_matrix, cval=255.0)
-    return sheared_im
-
-
-def vertical_shift(im, mode='mid'):
-    total = args.vertical_shift
-    if mode == 'mid':
-        top = int(total // 2)
-        bottom = total - top
-    elif mode == 'top':  # more padding on top
-        top = random.randint(total / 2, total)
-        bottom = total - top
-    elif mode == 'bottom':  # more padding on bottom
-        top = random.randint(0, total / 2)
-        bottom = total - top
-    width = im.shape[1]
-    im_pad = np.concatenate(
-        (255 * np.ones((top, width), dtype=int) -
-         np.random.normal(2, 1, (top, width)).astype(int), im), axis=0)
-    im_pad = np.concatenate(
-        (im_pad, 255 * np.ones((bottom, width), dtype=int) -
-         np.random.normal(2, 1, (bottom, width)).astype(int)), axis=0)
-    return im_pad
-
-
-# main #
-
-random.seed(1)
-
-scp_name = 'images.scp'  # parallel
-data_list_path = os.path.join(args.dir, scp_name)
-# output dir of feature matrix
 if args.out_ark == '-':
     out_fh = sys.stdout
 else:
-    out_fh = open(args.out_ark, 'wb')
-
+    out_fh = open(args.out_ark,'wb')
 
 allowed_lengths = None
 if os.path.isfile(os.path.join(args.dir, 'allowed_lengths.txt')):
@@ -263,19 +113,15 @@ with open(data_list_path) as f:
         image_id = line_vect[0]
         image_path = line_vect[1]
         im = misc.imread(image_path)
-        im_scaled = get_scaled_image(im)
-        im_contrast = contrast_normalization(im_scaled, 0.05, 0.2)
-        slant_degree = find_slant_project(im_contrast)
-        im_sheared = horizontal_shear(im_contrast, slant_degree)
-        im_horizontal_padded = horizontal_pad(im_sheared, allowed_lengths)
-        if im_horizontal_padded is None:
+        im_scaled = get_scaled_image(im, allowed_lengths)
+
+        if im_scaled is None:
             num_fail += 1
             continue
-        #im_padded = vertical_shift(im_horizontal_padded)
-        data = np.transpose(im_horizontal_padded, (1, 0))
+        data = np.transpose(im_scaled, (1, 0))
         data = np.divide(data, 255.0)
         num_ok += 1
         write_kaldi_matrix(out_fh, data, image_id)
 
-print('Generated features for {} images. Failed for {} (image too '
+print('Generated features for {} images. Failed for {} (iamge too '
       'long).'.format(num_ok, num_fail), file=sys.stderr)
